requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
argon2-cffi==20.1.0
    # via django
asgiref==3.3.4
    # via django
certifi==2020.12.5
    # via sentry-sdk
cffi==1.14.5
    # via argon2-cffi
dj-database-url==0.5.0
    # via -r requirements.in
django-csp==3.7
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-enforce-host==1.0.1
    # via -r requirements.in
django-permissions-policy==4.0.1
    # via -r requirements.in
django[argon2]==3.2.3
    # via
    #   -r requirements.in
    #   django-csp
    #   django-debug-toolbar
    #   django-permissions-policy
gunicorn==20.1.0
    # via -r requirements.in
psycopg2==2.8.6
    # via -r requirements.in
pycparser==2.20
    # via cffi
pytz==2021.1
    # via django
sentry-sdk==1.1.0
    # via -r requirements.in
six==1.16.0
    # via argon2-cffi
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
urllib3==1.26.4
    # via sentry-sdk
whitenoise==5.2.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
